Reverted commit #8de36d7
