+++ src/EdgeWARN/PreProcess/CellDetection/tracker.py
@@ -2,9 +2,9 @@ import json
 from .detect import detect_cells
 from pathlib import Path
 # assumes your existing imports: detect, load, match_cells, process_matched_cell, vectors, plot_radar_and_cells
-from util.PreProcess.data_utils import load_mrms_slice
-from util.PreProcess.save import StormCellTracker, save_cells_to_json
-from util.PreProcess.visualize import Visualizer
+from ..core.utils import load_mrms_slice
+from ..core.save import StormCellTracker, save_cells_to_json
+from ..core.visualize import Visualizer
 
 class StormCellDataManager:
     def __init__(self, storm_json: Path):
